+++ energy_view.cpp
@@ -448,11 +448,11 @@ void EnergyView::showPageFromTable()
 	Page::showPage();
 }
 
-QMap<int, float> EnergyView::convertGraphData(GraphData *gd)
+QMap<int, double> EnergyView::convertGraphData(GraphData *gd)
 {
 	// TODO: remove the theorical bottleneck using a QMutableMapIterator instead of
 	// accessing to the values using the operator[]
-	QMap<int, float> data;
+	QMap<int, double> data;
 	// copy map to float values
 	QList<int> graph_keys = gd->graph.keys();
 	for (int i = 0; i < graph_keys.size(); ++i)
@@ -527,7 +527,7 @@ void EnergyView::status_changed(const StatusList &status_list)
 			if (current_graph == EnergyDevice::DAILY_AVERAGE && date.year() == current_date.year() &&
 				date.month() == current_date.month() && isGraphOurs())
 			{
-				QMap<int, float> g = convertGraphData(d);
+				QMap<int, double> g = convertGraphData(d);
 				graph->setData(g);
 				table->setData(g);
 			}
@@ -539,7 +539,7 @@ void EnergyView::status_changed(const StatusList &status_list)
 			if (current_graph == EnergyDevice::CUMULATIVE_DAY && d->date == current_date &&
 			    isGraphOurs())
 			{
-				QMap<int, float> g = convertGraphData(d);
+				QMap<int, double> g = convertGraphData(d);
 				graph->setData(g);
 				table->setData(g);
 			}
@@ -551,7 +551,7 @@ void EnergyView::status_changed(const StatusList &status_list)
 			if (current_graph == EnergyDevice::CUMULATIVE_MONTH && d->date.month() == current_date.month()
 				&& d->date.year() == current_date.year() && isGraphOurs())
 			{
-				QMap<int, float> g = convertGraphData(d);
+				QMap<int, double> g = convertGraphData(d);
 				graph->setData(g);
 				table->setData(g);
 			}
@@ -563,7 +563,7 @@ void EnergyView::status_changed(const StatusList &status_list)
 			GraphData *d = saveGraphInCache(it.value(), EnergyDevice::CUMULATIVE_YEAR);
 			if (current_graph == EnergyDevice::CUMULATIVE_YEAR && isGraphOurs())
 			{
-				QMap<int, float> g = convertGraphData(d);
+				QMap<int, double> g = convertGraphData(d);
 				graph->setData(g);
 				table->setData(g);
 			}
@@ -630,7 +630,7 @@ void EnergyView::updateCurrentGraph()
 	if (graph_data_cache.contains(current_graph) && graph_data_cache[current_graph]->contains(key))
 	{
 		GraphData *d = graph_data_cache[current_graph]->object(key);
-		QMap<int, float> g = convertGraphData(d);
+		QMap<int, double> g = convertGraphData(d);
 		graph->setData(g);
 		table->setData(g);
 	}
